flask_sample/requirements.txt
attrs==22.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blinker==1.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
filelock==3.8.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-MySQLdb==2.0.0
Flask-Principal==0.4.0
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.2.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
requests==2.31.0
soupsieve==2.3.2.post1
tomli==2.0.1
typish==1.9.3
urllib3==2.1.0
virtualenv==20.16.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
